+++ lib/smart_core/types/primitive.rb
@@ -2,16 +2,19 @@
 
 # @api private
 # @since 0.1.0
+# @version 0.2.0
 class SmartCore::Types::Primitive
+  require_relative 'primitive/checker'
   require_relative 'primitive/caster'
   require_relative 'primitive/undefined_caster'
-  require_relative 'primitive/checker'
-  require_relative 'primitive/nilable_checker'
-  require_relative 'primitive/sum_checker'
-  require_relative 'primitive/mult_checker'
+  require_relative 'primitive/invariant_control'
+  require_relative 'primitive/validator'
   require_relative 'primitive/factory'
+  require_relative 'primitive/sum_validator'
   require_relative 'primitive/sum_factory'
+  require_relative 'primitive/mult_validator'
   require_relative 'primitive/mult_factory'
+  require_relative 'primitive/nilable_validator'
   require_relative 'primitive/nilable_factory'
 
   class << self
@@ -26,11 +29,12 @@ class SmartCore::Types::Primitive
     end
   end
 
-  # @return [SmartCore::Types::Primitive::Checker]
+  # @note NilClass is suitable for sum-types, mult-types and nilable types.
+  # @return [String, NilClass]
   #
-  # @api private
-  # @since 0.1.0
-  attr_reader :checker
+  # @api public
+  # @since 0.2.0
+  attr_reader :name
 
   # @return [SmartCore::Types::Primitive::Caster]
   #
@@ -38,17 +42,34 @@ class SmartCore::Types::Primitive
   # @since 0.1.0
   attr_reader :caster
 
-  # @param checker [SmartCore::Types::Primitive::Checker]
+  # @return [SmartCore::Types::Primitive::Validator]
+  # @return [SmartCore::Types::Primitive::SumValidator]
+  # @return [SmartCore::Types::Primitive::MultValidator]
+  # @return [SmartCore::Types::primitive::NilableValidator]
+  #
+  # @api private
+  # @since 0.2.0
+  attr_reader :validator
+
+  # @param name [String, NilClass] NilClass is suitable for sum-types, mult-types and nilable types.
+  # @param validator [
+  #   SmartCore::Types::Primitive::Validator,
+  #   SmartCore::Types::Primitive::SumValidator,
+  #   SmartCore::Types::Primitive::MultValidator,
+  #   SmartCore::Types::Primitive::NilableValidator
+  # ]
   # @param caster [SmartCore::Types::Primitive::Caster]
   # @return [void]
   #
   # @api private
   # @since 0.1.0
-  def initialize(checker, caster)
-    @lock = SmartCore::Engine::Lock.new
-    @checker = checker
+  # @version 0.2.0
+  def initialize(name, validator, caster)
+    @name = name
+    @validator = validator
     @caster = caster
     @nilable = nil
+    @lock = SmartCore::Engine::Lock.new
   end
 
   # @param value [Any]
@@ -56,20 +77,38 @@ class SmartCore::Types::Primitive
   #
   # @api public
   # @since 0.1.0
+  # @since 0.2.0
   def valid?(value)
-    checker.call(value)
+    validator.valid?(value)
   end
 
   # @param value [Any]
   # @return [void]
   #
   # @raise [SmartCore::Types::TypeError]
+  # @see SmartCore::Primitive::Validator
+  # @see SmartCore::Primitive::MultValidator
+  # @see SmartCore::Primitive::SumValidator
+  # @see SmartCore::Primitive::NilableValidator
   #
   # @api public
   # @since 0.1.0
+  # @version 0.2.0
   def validate!(value)
-    return if valid?(value)
-    raise(SmartCore::Types::TypeError, 'Invalid type')
+    validator.validate!(value)
+  end
+
+  # @return [SmartCore::Types::Primitive::Validator::Result]
+  #
+  # @see SmartCore::Primitive::Validator
+  # @see SmartCore::Primitive::MultValidator
+  # @see SmartCore::Primitive::SumValidator
+  # @see SmartCore::Primitive::NilableValidator
+  #
+  # @api public
+  # @since 0.2.0
+  def validate(value)
+    validator.validate(value)
   end
 
   # @param value [Any]