+++ lib/smart_core/types/primitive/mult_factory/definition_context.rb
@@ -2,7 +2,7 @@
 
 # @api private
 # @since 0.1.0
-class SmartCore::Types::Primitive::Factory::DefinitionContext
+class SmartCore::Types::Primitive::MultFactory::DefinitionContext
   # @return [Proc, NilClass]
   #
   # @api private
@@ -23,7 +23,9 @@ class SmartCore::Types::Primitive::Factory::DefinitionContext
   # @api public
   # @since 0.1.0
   def define_caster(&caster)
-    raise(SmartCore::ArgumentError, 'No caster definition block') unless block_given?
+    raise(SmartCore::Types::ArgumentError, 'No caster definition block') unless block_given?
     @type_caster = caster
   end
+
+  # TODO (0.x.0): invariant API
 end