+++ lib/smart_core/types/primitive/mult_validator/result.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class SmartCore::Types::Primitive::MultValidator
+  # @api private
+  # @since 0.2.0
+  class Result < SmartCore::Types::Primitive::SumValidator::Result
+  end
+end