+++ lib/smart_core/types/primitive/nilable_factory.rb
@@ -2,6 +2,7 @@
 
 # @api private
 # @since 0.1.0
+# @version 0.2.0
 module SmartCore::Types::Primitive::NilableFactory
   class << self
     # @param type [SmartCore::Types::Primitive]
@@ -9,21 +10,24 @@ module SmartCore::Types::Primitive::NilableFactory
     #
     # @api private
     # @since 0.1.0
+    # @version 0.2.0
     def create_type(type)
-      type_checker = build_type_checker(type)
+      type_validator = build_type_validator(type)
       type_caster = build_type_caster(type)
-      build_type(type, type_checker, type_caster)
+      build_type(type, type_validator, type_caster).tap do |new_type|
+        assign_type_validator(new_type, type_validator)
+      end
     end
 
     private
 
     # @param type [SmartCore::Types::Primitive]
-    # @return [SmartCore::Types::Primitive::NilableChecker]
+    # @return [SmartCore::Types::Primitive::NilableValidator]
     #
     # @api private
-    # @since 0.1.0
-    def build_type_checker(type)
-      SmartCore::Types::Primitive::NilableChecker.new(type.checker)
+    # @since 0.2.0
+    def build_type_validator(type)
+      SmartCore::Types::Primitive::NilableValidator.new(type.validator)
     end
 
     # @param type [SmartCore::Types::Primitive]
@@ -36,14 +40,25 @@ module SmartCore::Types::Primitive::NilableFactory
     end
 
     # @param type [SmartCore::Types::Primitive]
-    # @param type_checker [SmartCore::Types::Primitive::NilableChecker]
+    # @param type_validator [SmartCore::Types::Primitive::NilableValidator]
+    # @return [void]
+    #
+    # @api private
+    # @since 0.2.0
+    def assign_type_validator(type, type_validator)
+      type_validator.___assign_type___(type)
+    end
+
+    # @param type [SmartCore::Types::Primitive]
+    # @param type_validator [SmartCore::Types::Primitive::NilableValidator]
     # @param type_caster [SmartCore::Types::Caster]
     # @return [SmartCore::Type::Primitive]
     #
     # @api private
     # @since 0.1.0
-    def build_type(type, type_checker, type_caster)
-      Class.new(type.class).new(type_checker, type_caster)
+    # @version 0.2.0
+    def build_type(type, type_validator, type_caster)
+      Class.new(type.class).new(type.name, type_validator, type_caster)
     end
   end
 end