+++ lib/smart_core/types/value/nil.rb
@@ -4,8 +4,6 @@
 # @since 0.1.0
 SmartCore::Types::Value.define_type(:Nil) do |type|
   type.define_checker do |value|
-    # rubocop:disable Style/NilComparison
     value == nil # NOTE: #nil? is not used cuz BasicObject hasn't #nil? method
-    # rubocop:enable Style/NilComparison
   end
 end