+++ lib/smart_core/types/value/range.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # @api public
-# @since 0.2.0
+# @since 0.x.0
 SmartCore::Types::Value.define_type(:Range) do |type|
   type.define_checker do |value|
     # TODO: realize