+++ spec/spec_helper.rb
@@ -6,10 +6,12 @@ SimpleCov.formatter = SimpleCov::Formatter::HTMLFormatter
 SimpleCov.minimum_coverage(100)
 SimpleCov.start do
   enable_coverage :branch
+  enable_coverage :line
   add_filter 'spec'
 end
 
 require 'bundler/setup'
+require 'pry'
 require 'smart_core/types'
 
 RSpec.configure do |config|