+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Added
+- Brand new **Type invariant API**:
+  - globally refactored validation logic (with backward compatability for `#valid?(value)` method);
+  - new type definition DSL: `.invariant(name)` and `.invariant_chain(name)`;
+  - chained invariants will be invoked according to the definition order (second invokation
+    depends on previous successful invariant check);
+  - new validation API: `validate(value)` (with `#errors` support based on invariant names);
+  - at this moment Invariant API is supported only by primitive types (type sum and type multiplication support coming soon);
+
+### Changed
+
+- Updated `smart_engine` dependency (to `~> 0.7`) (need `SmartCore::Engine::Atom`);
+
 ## [0.1.0] - 2020-05-05
 - Release :)